+++ test/fixtures/test_dependency.ex
@@ -1,3 +1,0 @@
-defmodule Test.Dependency do
-  def call, do: "Test.Dependency"  
-end