+++ test/fixtures/test_module.ex
@@ -1,3 +1,0 @@
-defmodule Test.TestModule do
-  def call, do: Test.Dependency.call()
-end