+++ CHANGELOG.md
@@ -1,4 +1,14 @@
+## [1.1.0] - 2017.09.25
+
+- no func decorator
+- no separate Container module
+- issue with args calculation was fixed
+- `when` option (opts - keywords list, 3d argument in `swap/3`)
+- raises an error in a case when function wasnt defined in a 'swapper' module but was invoked
+- tests were simplified
+
 ## [1.0.1] - 2017.09.25
+
 - arjan/decorator lib has been updated
 
 ## [1.0.0] - 2017.08.30